# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ساعدني - API specification for the Oman help marketplace
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
}

export type UserUserType = typeof UserUserType[keyof typeof UserUserType];


export const UserUserType = {
  customer: 'customer',
  helper: 'helper',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  phone: string;
  userType: UserUserType;
  /** @nullable */
  area?: string | null;
  /** @nullable */
  rating?: number | null;
  isVerified?: boolean;
  isBlocked?: boolean;
  createdAt: string;
}

export type RegisterInputUserType = typeof RegisterInputUserType[keyof typeof RegisterInputUserType];


export const RegisterInputUserType = {
  customer: 'customer',
  helper: 'helper',
} as const;

export interface RegisterInput {
  name: string;
  phone: string;
  password: string;
  userType: RegisterInputUserType;
  area?: string;
}

export interface LoginInput {
  phone: string;
  password: string;
}

export interface AuthResponse {
  user: User;
}

export interface UserUpdate {
  name?: string;
  area?: string;
}

export type HelpRequestCategory = typeof HelpRequestCategory[keyof typeof HelpRequestCategory];


export const HelpRequestCategory = {
  transport: 'transport',
  delivery: 'delivery',
  government: 'government',
  shopping: 'shopping',
  home_services: 'home_services',
  labor: 'labor',
} as const;

export type HelpRequestTimeType = typeof HelpRequestTimeType[keyof typeof HelpRequestTimeType];


export const HelpRequestTimeType = {
  now: 'now',
  scheduled: 'scheduled',
} as const;

export type HelpRequestStatus = typeof HelpRequestStatus[keyof typeof HelpRequestStatus];


export const HelpRequestStatus = {
  available: 'available',
  accepted: 'accepted',
  in_progress: 'in_progress',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface HelpRequest {
  id: number;
  customerId: number;
  /** @nullable */
  helperId?: number | null;
  category: HelpRequestCategory;
  details: string;
  area: string;
  timeType: HelpRequestTimeType;
  /** @nullable */
  scheduledDateTime?: string | null;
  offeredAmount: number;
  status: HelpRequestStatus;
  createdAt: string;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  helperName?: string | null;
}

export type HelpRequestInputCategory = typeof HelpRequestInputCategory[keyof typeof HelpRequestInputCategory];


export const HelpRequestInputCategory = {
  transport: 'transport',
  delivery: 'delivery',
  government: 'government',
  shopping: 'shopping',
  home_services: 'home_services',
  labor: 'labor',
} as const;

export type HelpRequestInputTimeType = typeof HelpRequestInputTimeType[keyof typeof HelpRequestInputTimeType];


export const HelpRequestInputTimeType = {
  now: 'now',
  scheduled: 'scheduled',
} as const;

export interface HelpRequestInput {
  customerId: number;
  category: HelpRequestInputCategory;
  details: string;
  area: string;
  timeType: HelpRequestInputTimeType;
  scheduledDateTime?: string;
  offeredAmount: number;
}

export interface HelpRequestUpdate {
  category?: string;
  details?: string;
  area?: string;
  timeType?: string;
  scheduledDateTime?: string;
  offeredAmount?: number;
  status?: string;
}

export interface AcceptRequestInput {
  helperId: number;
}

export interface AdminStats {
  totalUsers: number;
  totalHelpers: number;
  totalCustomers: number;
  totalRequests: number;
  activeRequests: number;
  completedRequests: number;
  cancelledRequests: number;
}

export type VerifyHelperInputAction = typeof VerifyHelperInputAction[keyof typeof VerifyHelperInputAction];


export const VerifyHelperInputAction = {
  verify: 'verify',
  block: 'block',
} as const;

export interface VerifyHelperInput {
  action: VerifyHelperInputAction;
}

export type ListRequestsParams = {
category?: string;
area?: string;
status?: string;
customerId?: number;
};

export type ListUsersParams = {
userType?: string;
};