# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ساعدني - API specification for the Oman help marketplace
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: users
    description: User operations
  - name: requests
    description: Help request operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List help requests
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: area
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HelpRequest"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Create a new help request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HelpRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelpRequest"
        "400":
          description: Bad request

  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get a specific help request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelpRequest"
        "404":
          description: Not found
    patch:
      operationId: updateRequest
      tags: [requests]
      summary: Update a help request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HelpRequestUpdate"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelpRequest"
        "404":
          description: Not found
    delete:
      operationId: deleteRequest
      tags: [requests]
      summary: Delete a help request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /requests/{id}/accept:
    patch:
      operationId: acceptRequest
      tags: [requests]
      summary: Helper accepts a request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptRequestInput"
      responses:
        "200":
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelpRequest"
        "400":
          description: Already accepted
        "404":
          description: Not found

  /requests/{id}/cancel:
    patch:
      operationId: cancelRequest
      tags: [requests]
      summary: Customer cancels a request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelpRequest"
        "404":
          description: Not found

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin)
      parameters:
        - name: userType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Admin statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /admin/helpers/{id}/verify:
    patch:
      operationId: verifyHelper
      tags: [admin]
      summary: Approve or block a helper
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyHelperInput"
      responses:
        "200":
          description: Helper updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    User:
      type: object
      required: [id, name, phone, userType, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        userType:
          type: string
          enum: [customer, helper, admin]
        area:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        isVerified:
          type: boolean
        isBlocked:
          type: boolean
        createdAt:
          type: string

    RegisterInput:
      type: object
      required: [name, phone, password, userType]
      properties:
        name:
          type: string
        phone:
          type: string
        password:
          type: string
        userType:
          type: string
          enum: [customer, helper]
        area:
          type: string

    LoginInput:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string

    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    UserUpdate:
      type: object
      properties:
        name:
          type: string
        area:
          type: string

    HelpRequest:
      type: object
      required: [id, customerId, category, details, area, timeType, offeredAmount, status, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        helperId:
          type: ["integer", "null"]
        category:
          type: string
          enum: [transport, delivery, government, shopping, home_services, labor]
        details:
          type: string
        area:
          type: string
        timeType:
          type: string
          enum: [now, scheduled]
        scheduledDateTime:
          type: ["string", "null"]
        offeredAmount:
          type: number
        status:
          type: string
          enum: [available, accepted, in_progress, completed, cancelled]
        createdAt:
          type: string
        customerName:
          type: ["string", "null"]
        helperName:
          type: ["string", "null"]

    HelpRequestInput:
      type: object
      required: [customerId, category, details, area, timeType, offeredAmount]
      properties:
        customerId:
          type: integer
        category:
          type: string
          enum: [transport, delivery, government, shopping, home_services, labor]
        details:
          type: string
        area:
          type: string
        timeType:
          type: string
          enum: [now, scheduled]
        scheduledDateTime:
          type: string
        offeredAmount:
          type: number

    HelpRequestUpdate:
      type: object
      properties:
        category:
          type: string
        details:
          type: string
        area:
          type: string
        timeType:
          type: string
        scheduledDateTime:
          type: string
        offeredAmount:
          type: number
        status:
          type: string

    AcceptRequestInput:
      type: object
      required: [helperId]
      properties:
        helperId:
          type: integer

    AdminStats:
      type: object
      required: [totalUsers, totalHelpers, totalCustomers, totalRequests, activeRequests, completedRequests, cancelledRequests]
      properties:
        totalUsers:
          type: integer
        totalHelpers:
          type: integer
        totalCustomers:
          type: integer
        totalRequests:
          type: integer
        activeRequests:
          type: integer
        completedRequests:
          type: integer
        cancelledRequests:
          type: integer

    VerifyHelperInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [verify, block]