# lib/api-zod/src/generated/types/acceptRequestInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ساعدني - API specification for the Oman help marketplace
 * OpenAPI spec version: 0.1.0
 */

export interface AcceptRequestInput {
  helperId: number;
}