# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ساعدني - API specification for the Oman help marketplace
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalUsers: number;
  totalHelpers: number;
  totalCustomers: number;
  totalRequests: number;
  activeRequests: number;
  completedRequests: number;
  cancelledRequests: number;
}