# lib/api-zod/src/generated/types/helpRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ساعدني - API specification for the Oman help marketplace
 * OpenAPI spec version: 0.1.0
 */
import type { HelpRequestCategory } from './helpRequestCategory';
import type { HelpRequestStatus } from './helpRequestStatus';
import type { HelpRequestTimeType } from './helpRequestTimeType';

export interface HelpRequest {
  id: number;
  customerId: number;
  /** @nullable */
  helperId?: number | null;
  category: HelpRequestCategory;
  details: string;
  area: string;
  timeType: HelpRequestTimeType;
  /** @nullable */
  scheduledDateTime?: string | null;
  offeredAmount: number;
  status: HelpRequestStatus;
  createdAt: string;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  helperName?: string | null;
}