# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ساعدني - API specification for the Oman help marketplace
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterInputUserType } from './registerInputUserType';

export interface RegisterInput {
  name: string;
  phone: string;
  password: string;
  userType: RegisterInputUserType;
  area?: string;
}