# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ساعدني - API specification for the Oman help marketplace
 * OpenAPI spec version: 0.1.0
 */
import type { UserUserType } from './userUserType';

export interface User {
  id: number;
  name: string;
  phone: string;
  userType: UserUserType;
  /** @nullable */
  area?: string | null;
  /** @nullable */
  rating?: number | null;
  isVerified?: boolean;
  isBlocked?: boolean;
  createdAt: string;
}