# lib/api-zod/src/generated/types/verifyHelperInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ساعدني - API specification for the Oman help marketplace
 * OpenAPI spec version: 0.1.0
 */
import type { VerifyHelperInputAction } from './verifyHelperInputAction';

export interface VerifyHelperInput {
  action: VerifyHelperInputAction;
}